First commit
This commit adds basic friend tracking functionality to slither.io.
It does this by opening up a socket.io connection with the given server
and sends periodic coord updates to the server. The server also sends
all connected player data to the client so that it can draw icons on the minimap.
These icons are green if it is a friend, and red if it is a dead friend.